Update 2014-08-20-seven-questions-part-2.markdown
Yo. 

I was having a hard time following the article to be honest -- how themes/points were tying together. Let me know what you think of the direction I took it. Definitely it needs edits to be in your voice, and the bullets under "on motives" need work, but I gotta eat a sandwich right now!! And didn't want to hold you up.  :)